Go live: the bubble journey becomes the homepage
- Extract the option R scene into src/components/home/BubbleHome.tsx
- Index renders BubbleHome with the live canonical and meta (indexable)
- /home-preview-r stays as the noindex design-lane wrapper + ribbon
- Add Index.test.tsx: signature heading, install path, plain language,
  and a guard that the live page never carries noindex
